feat: Add LanguageToggle, ModeToggle, and ThemeProvider components
- Implement LanguageToggle component for switching between languages.
- Create ModeToggle component for toggling between light and dark themes.
- Introduce ThemeProvider for managing theme state and persistence.
- Add TestComponent for testing user data fetching and updating.
- Develop AppSidebar and MobileHeader for navigation and user interaction.
- Create CSS modules for styling components.
- Define TypeScript interfaces for common props and types.
