+++ src/app/(auth)/login/page.tsx
@@ -1,4 +1,4 @@
-import { AuthPage } from "@/components/auth-page"
+import { AuthPage } from "@/components/auth"
 
 export default function LoginPage() {
   return <AuthPage />
