+++ src/app/(auth)/register/page.tsx
@@ -1,4 +1,4 @@
-import { AuthPage } from "@/components/auth-page"
+import { AuthPage } from "@/components/auth"
 
 export default function RegisterPage() {
   return <AuthPage />
