+++ src/app/@auth/(.)register/page.tsx
@@ -1,4 +1,4 @@
-import { AuthModal } from "@/components/auth-modal"
+import { AuthModal } from "@/components/auth"
 
 export default function RegisterModal() {
   return <AuthModal />
