--- src/components/auth-modal.tsx
@@ -1,32 +1,0 @@
-"use client"
-
-import { useState, useEffect } from "react"
-import { useRouter, usePathname } from "next/navigation"
-import { Dialog, DialogContent } from "@/components/ui/dialog"
-import { AuthForm } from "@/components/auth-form"
-
-export function AuthModal() {
-  const router = useRouter()
-  const pathname = usePathname()
-  const [isOpen, setIsOpen] = useState(true)
-  const [mode, setMode] = useState<"login" | "register">(pathname.includes("register") ? "register" : "login")
-
-  useEffect(() => {
-    setMode(pathname.includes("register") ? "register" : "login")
-  }, [pathname])
-
-  const handleClose = () => {
-    setIsOpen(false)
-    setTimeout(() => router.back(), 150)
-  }
-
-  return (
-    <Dialog open={isOpen} onOpenChange={handleClose}>
-      <DialogContent className="sm:max-w-md p-0 overflow-hidden border-0 bg-transparent shadow-2xl">
-        <div className="bg-background rounded-lg">
-          <AuthForm mode={mode} onModeChange={setMode} isModal />
-        </div>
-      </DialogContent>
-    </Dialog>
-  )
-}
