+++ src/components/auth/AuthForm.module.css
@@ -1,0 +1,1 @@
+/* AuthForm specific styles - only for custom CSS that can't be handled by Tailwind */
