+++ src/components/auth/AuthModal.module.css
@@ -1,0 +1,1 @@
+/* AuthModal specific styles - only for custom CSS that can't be handled by Tailwind */
