+++ src/components/auth/AuthPage.module.css
@@ -1,0 +1,1 @@
+/* AuthPage specific styles - only for custom CSS that can't be handled by Tailwind */
