+++ src/components/auth/index.ts
@@ -1,0 +1,3 @@
+export { AuthForm } from './AuthForm'
+export { AuthModal } from './AuthModal'
+export { AuthPage } from './AuthPage'
