+++ src/components/chat/ChatDock.module.css
@@ -1,0 +1,1 @@
+/* ChatDock specific styles - only for custom CSS that can't be handled by Tailwind */
