+++ src/components/chat/index.ts
@@ -1,0 +1,1 @@
+export { ChatDock } from './ChatDock'
