+++ src/components/common/LanguageToggle.module.css
@@ -1,0 +1,1 @@
+/* LanguageToggle specific styles - only for custom CSS that can't be handled by Tailwind */
