+++ src/components/common/ModeToggle.module.css
@@ -1,0 +1,1 @@
+/* ModeToggle specific styles - only for custom CSS that can't be handled by Tailwind */
