+++ src/components/common/TestComponent.module.css
@@ -1,0 +1,1 @@
+/* TestComponent specific styles - only for custom CSS that can't be handled by Tailwind */
