+++ src/components/common/ThemeProvider.module.css
@@ -1,0 +1,1 @@
+/* ThemeProvider specific styles - only for custom CSS that can't be handled by Tailwind */
