--- src/components/language-toggle.tsx
@@ -1,42 +1,0 @@
-"use client"
-
-import { Button } from "@/components/ui/button"
-import { useLanguage } from "@/lib/language-context"
-import { Languages } from "lucide-react"
-
-interface LanguageToggleProps {
-  collapsed?: boolean
-}
-
-export function LanguageToggle({ collapsed = false }: LanguageToggleProps) {
-  const { language, setLanguage } = useLanguage()
-
-  const toggleLanguage = () => {
-    setLanguage(language === "en" ? "vi" : "en")
-  }
-
-  if (collapsed) {
-    return (
-      <Button
-        variant="ghost"
-        size="icon"
-        onClick={toggleLanguage}
-        className="h-9 w-9 hover:bg-orange-100 dark:hover:bg-orange-900/20"
-        title={language === "en" ? "Switch to Vietnamese" : "Chuyển sang tiếng Anh"}
-      >
-        <Languages className="h-4 w-4" />
-      </Button>
-    )
-  }
-
-  return (
-    <Button
-      variant="ghost"
-      onClick={toggleLanguage}
-      className="w-full justify-start gap-2 hover:bg-orange-100 dark:hover:bg-orange-900/20"
-    >
-      <Languages className="h-4 w-4" />
-      <span className="text-sm font-medium">{language === "en" ? "VI" : "EN"}</span>
-    </Button>
-  )
-}
