+++ src/components/layout/AppSidebar.module.css
@@ -1,0 +1,1 @@
+/* AppSidebar specific styles - only for custom CSS that can't be handled by Tailwind */
