+++ src/components/layout/MobileHeader.module.css
@@ -1,0 +1,1 @@
+/* MobileHeader specific styles - only for custom CSS that can't be handled by Tailwind */
