--- src/components/theme-provider.tsx
@@ -1,84 +1,0 @@
-"use client"
-
-import * as React from "react"
-
-type Theme = "dark" | "light" | "system"
-
-interface ThemeProviderProps {
-  children: React.ReactNode
-  attribute?: string
-  defaultTheme?: Theme
-  enableSystem?: boolean
-  disableTransitionOnChange?: boolean
-}
-
-interface ThemeProviderState {
-  theme: Theme
-  setTheme: (theme: Theme) => void
-}
-
-const ThemeProviderContext = React.createContext<ThemeProviderState | undefined>(undefined)
-
-export function ThemeProvider({
-  children,
-  attribute = "class",
-  defaultTheme = "system",
-  enableSystem = true,
-  disableTransitionOnChange = false,
-  ...props
-}: ThemeProviderProps) {
-  const [theme, setThemeState] = React.useState<Theme>(defaultTheme)
-
-  React.useEffect(() => {
-    const stored = localStorage.getItem("theme") as Theme
-    if (stored) {
-      setThemeState(stored)
-    } else if (enableSystem) {
-      const systemTheme = window.matchMedia("(prefers-color-scheme: dark)").matches ? "dark" : "light"
-      setThemeState(systemTheme)
-    }
-  }, [enableSystem])
-
-  React.useEffect(() => {
-    const root = window.document.documentElement
-
-    if (disableTransitionOnChange) {
-      root.classList.add("[&_*]:!transition-none")
-      window.setTimeout(() => {
-        root.classList.remove("[&_*]:!transition-none")
-      }, 1)
-    }
-
-    root.classList.remove("light", "dark")
-
-    if (theme === "system" && enableSystem) {
-      const systemTheme = window.matchMedia("(prefers-color-scheme: dark)").matches ? "dark" : "light"
-      root.classList.add(systemTheme)
-    } else {
-      root.classList.add(theme)
-    }
-  }, [theme, enableSystem, disableTransitionOnChange])
-
-  const setTheme = React.useCallback((newTheme: Theme) => {
-    localStorage.setItem("theme", newTheme)
-    setThemeState(newTheme)
-  }, [])
-
-  const value = React.useMemo(
-    () => ({
-      theme,
-      setTheme,
-    }),
-    [theme, setTheme],
-  )
-
-  return <ThemeProviderContext.Provider value={value}>{children}</ThemeProviderContext.Provider>
-}
-
-export const useTheme = () => {
-  const context = React.useContext(ThemeProviderContext)
-
-  if (context === undefined) throw new Error("useTheme must be used within a ThemeProvider")
-
-  return context
-}
